Switched to loading the whole job in earlier
Can lead to a race condition. What if the DistrictingSingleton isn't finished setting up by the time the user wants to constrain?

+++ spring/src/main/java/cse416/spring/service/DistrictingService.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public interface DistrictingService {
     Districting findById(long id);
 
-    List<Districting> findByJob(int jobId);
+    List<Districting> findByJob(long jobId);
 
     List<Districting> findAllDistrictings();
